+++ lib/src/app_initializer.dart
@@ -1,6 +1,8 @@
 import 'package:components/components.dart';
 import 'package:core/core.dart';
 import 'package:core_widgets/core_widgets.dart';
+import 'package:flutter_trc/src/libraries/firebase/firebase_helper.dart';
+import 'package:flutter_trc/src/libraries/firebase/remote_config_helper.dart';
 import 'package:localization/localization.dart';
 
 import 'actions/project_actions.dart';
@@ -18,6 +20,8 @@ const RUNNING_SYSTEM_ENV = String.fromEnvironment('env', defaultValue: 'prod');
 class AppInitializer {
   static init({required List<AnalyticTrackers> trackers, Map<String, HttpInterceptorFactory>? interceptors}) async {
     await AuthHandler().syncAuth();
+    await FirebaseHelper().initFirebase();
+    await RemoteConfigHelper().initialize();
     await AnalyticsController.init(trackers);
     await _initApp(interceptors: interceptors);
   }