+++ lib/src/libraries/firebase/firebase_helper.dart
@@ -1,0 +1,16 @@
+import 'package:firebase_core/firebase_core.dart';
+
+class FirebaseHelper {
+  static final FirebaseHelper _instance = FirebaseHelper._internal();
+
+  factory FirebaseHelper() {
+    return _instance;
+  }
+
+  FirebaseHelper._internal();
+
+  Future<FirebaseApp> initFirebase() async {
+    FirebaseApp firebaseApp = await Firebase.initializeApp();
+    return firebaseApp;
+  }
+}