+++ lib/src/utils/media_upload/media_optimiser_utils.dart
@@ -84,6 +84,7 @@ class MediaUploadUtil {
                     completer.completeError(error);
                   },
                   onSuccess: () {
+                    Logger.debug('mydebug-----MediaUploadUtil.uploadMedia---url-----', [_uploadedImageUrl]);
                     completer.complete(_uploadedImageUrl ?? "");
                   },
                 );